+++ src/main/java/com/pmn/gmt/domain/map/presentation/controller/MapController.java
@@ -2,8 +2,8 @@ package com.pmn.gmt.domain.map.presentation.controller;
 
 import com.pmn.gmt.domain.map.presentation.data.responseDto.MapResponseDto;
 import com.pmn.gmt.domain.map.presentation.data.responseDto.TeacherResponseDto;
-import com.pmn.gmt.domain.map.service.FindClassroomByIdService;
-import com.pmn.gmt.domain.map.service.FindTeachersByMapIdService;
+import com.pmn.gmt.domain.map.service.FindClassByNameService;
+import com.pmn.gmt.domain.map.service.FindTeachersByClassNameService;
 import com.pmn.gmt.domain.map.util.MapConverter;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -20,19 +20,19 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class MapController {
 
-    private final FindTeachersByMapIdService findTeachersByMapId;
+    private final FindTeachersByClassNameService findTeachersByClassNameService;
 
-    private final FindClassroomByIdService findClassroomByIdService;
+    private final FindClassByNameService findClassByNameService;
     private final MapConverter mapConverter;
 
     @GetMapping("/{class_name}")
-    ResponseEntity<MapResponseDto> findClassroomById(@PathVariable("class_name") String className){
-        return ResponseEntity.ok(mapConverter.toResponseDto(findClassroomByIdService.execute(mapConverter.toDto(className))));
+    ResponseEntity<MapResponseDto> findClassByName(@PathVariable("class_name") String className){
+        return ResponseEntity.ok(mapConverter.toResponseDto(findClassByNameService.execute(mapConverter.toDto(className))));
     }
 
     @GetMapping("/{class_name}/teachers")
-    ResponseEntity<List<TeacherResponseDto>> findTeachersByMapId(@PathVariable("class_name") String className){
-        return ResponseEntity.ok((findTeachersByMapId.execute(mapConverter.toDto(className))).stream()
+    ResponseEntity<List<TeacherResponseDto>> findTeachersByClassName(@PathVariable("class_name") String className){
+        return ResponseEntity.ok((findTeachersByClassNameService.execute(mapConverter.toDto(className))).stream()
                 .map(mapConverter::toResponseDto)
                 .collect(Collectors.toList()));
     }
