+++ src/main/java/com/pmn/gmt/domain/map/service/FindClassByNameService.java
@@ -3,6 +3,6 @@ package com.pmn.gmt.domain.map.service;
 import com.pmn.gmt.domain.map.presentation.data.dto.MapDto;
 import com.pmn.gmt.domain.map.presentation.data.dto.ClassNameDto;
 
-public interface FindClassroomByIdService {
+public interface FindClassByNameService {
     MapDto execute(ClassNameDto classNameDto);
 }
