+++ src/main/java/com/pmn/gmt/domain/map/service/impl/FindTeacherByClassNameServiceImpl.java
@@ -2,26 +2,31 @@ package com.pmn.gmt.domain.map.service.impl;
 
 import com.pmn.gmt.domain.map.domain.repository.MapRepository;
 import com.pmn.gmt.domain.map.exception.ClassRoomNotFoundException;
-import com.pmn.gmt.domain.map.presentation.data.dto.MapDto;
 import com.pmn.gmt.domain.map.presentation.data.dto.ClassNameDto;
-import com.pmn.gmt.domain.map.service.FindClassroomByIdService;
+import com.pmn.gmt.domain.map.presentation.data.dto.TeacherDto;
+import com.pmn.gmt.domain.map.service.FindTeachersByClassNameService;
 import com.pmn.gmt.domain.map.util.MapConverter;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 @Service
 @AllArgsConstructor
-public class FindClassroomByIdServiceImpl implements FindClassroomByIdService {
+public class FindTeacherByClassNameServiceImpl implements FindTeachersByClassNameService {
 
     private final MapRepository mapRepository;
     private final MapConverter mapConverter;
 
     @Override
-    public MapDto execute(ClassNameDto classNameDto) {
+    public List<TeacherDto> execute(ClassNameDto classNameDto) {
         if(mapRepository.findByName(classNameDto.getClassName()).isEmpty())
             throw new ClassRoomNotFoundException();
         else{
-            return mapConverter.toDto(mapRepository.findByName(classNameDto.getClassName()).get(0));
+            return mapRepository.findByName(classNameDto.getClassName()).stream()
+                    .map(it -> mapConverter.toDto(it.getTeacher()))
+                    .collect(Collectors.toList());
         }
     }
 }
